[OPTIMIZATION]: Abstraction in FeaturedSlider in separated Components and Hooks to apply Single Responsability
