+++ src/components/FeaturedSlider/SliderButton.jsx
@@ -1,0 +1,10 @@
+/* eslint-disable react/prop-types */
+import React from 'react'
+
+export default function SliderButton ({ onClickFunction, buttonClassName, iClassName }) {
+  return (
+    <button className={buttonClassName} onClick={onClickFunction}>
+      <i className={iClassName}></i>
+    </button>
+  )
+}
