+++ src/components/FeaturedSlider/index.jsx
@@ -1,65 +1,29 @@
-/* eslint-disable no-unused-vars */
-import React, { useContext, useEffect, useRef, useState } from 'react'
+import React, { useContext, useState } from 'react'
 import { BannersContext } from '../../providers/BannersProvider'
+import Banner from './Banner'
+import SliderButton from './SliderButton'
 import './styles.scss'
+import { useSliderFeatures } from './useSliderFeatures'
 
 export default function FeaturedSlider () {
   const { data, isLoading } = useContext(BannersContext)
   const [index, setIndex] = useState(0)
-  const timeoutRef = useRef(null)
-
-  function resetTimeout () {
-    if (timeoutRef.current) {
-      clearInterval(timeoutRef.current)
-    }
-  }
-
-  const prevSlider = () => {
-    resetTimeout()
-    if (index === 0) {
-      setIndex(data.results.length - 1)
-      return null
-    }
-    setIndex((prevIndex) => prevIndex - 1)
-  }
-
-  const nextSlider = () => {
-    if (index === data.results.length - 1) {
-      setIndex(0)
-      return null
-    }
-    setIndex((prevIndex) => prevIndex + 1)
-  }
-
-  useEffect(() => {
-    resetTimeout()
-    if (!isLoading && Object.keys(data).length !== 0) {
-      timeoutRef.current = setInterval(() => {
-        nextSlider()
-      }, 5000)
-    }
-    return () => {
-      resetTimeout()
-    }
-  }, [isLoading, data, index])
+  const { prevSlider, nextSlider } = useSliderFeatures({ data, isLoading, index, setIndex })
 
   return (
     <div className='home--slider'>
-      <button className='home--slider-prevButton' onClick={prevSlider}>
-        <i className="fa-solid fa-angles-left"></i>
-      </button>
+      <SliderButton
+        onClickFunction={prevSlider}
+        buttonClassName='home--slider-prevButton'
+        iClassName='fa-solid fa-angles-left'
+      />
       { isLoading && (<p>Loading ...</p>) }
       {
         data && data.results && (
           data.results.map((banner, bannerIndex) => {
             if (bannerIndex === index) {
               return (
-                <div className='home--slider--container' key={banner.id}>
-                  <img className='home--slider--container-image' src={banner.data.main_image.url} alt={banner.data.main_image.alt} />
-                  <p>
-                    { banner.data.title }
-                  </p>
-                </div>
+                <Banner banner={banner} key={banner.id} />
               )
             }
 
@@ -67,10 +31,11 @@ export default function FeaturedSlider () {
           })
         )
       }
-
-      <button className='home--slider-nextButton' onClick={nextSlider}>
-        <i className="fa-solid fa-angles-right"></i>
-      </button>
+      <SliderButton
+        onClickFunction={nextSlider}
+        buttonClassName='home--slider-nextButton'
+        iClassName='fa-solid fa-angles-right'
+      />
     </div>
   )
 }
